+++ CHANGELOG.md
@@ -24,8 +24,10 @@
     ([#228](https://github.com/microsoft/ApplicationInsights-Python/pull/228))
 - Removing diagnostic logging from its module's logger.
     ([#225](https://github.com/microsoft/ApplicationInsights-Python/pull/225))
-- Configure logging instrumenation root
+- Add ability to specify logger for logging configuration
     ([#227](https://github.com/microsoft/ApplicationInsights-Python/pull/227))
+- Add metric configuration to distro api
+    ([#232](https://github.com/microsoft/ApplicationInsights-Python/pull/232))
 
 ## [1.0.0b8](https://github.com/microsoft/ApplicationInsights-Python/releases/tag/v1.0.0b8) - 2022-09-26